Events.
Added an event system to Red. Basic string-based IDs that can resolve
to optimized dispatcher-local keys are used to index a list of listener
functions, abstracted by IFunction, allowing any of the implementing
closures to be the target of an event. IEvent objects are the method of
data passing for event dispatch.