Moved to double-precision for sample ratios
When resampling long tracks, it’s a good idea to have at least 32 bits
of precision in the mantissa of the sample ratio/inverse index
multiplier so as to avoid time-domain aliasing problems to due with
limited precision. Ideally i would be using a fixed point type here,
but double is currently the best option.